requirements.txt
pymonocypher==3.1.0.0